refactor: move swap types to swap/types.ts
